requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.6.1

# Data processing
pandas==2.2.0
numpy==1.26.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
